Add files via upload
In this word count Java project, we aim to develop a program that analyses
subtitles from movies to count the occurrences of each word. The subtitles are
assumed to use only ASCII characters and include casual English, contractions,
and punctuation. The program will tokenize the subtitles by splitting them based
on whitespace and punctuation, treating contractions as single words.